requirements.txt
flask>=3.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
ollama>=0.1.0
